requirements.txt
requests
PyYAML
python-dotenv
beautifulsoup4
lxml
python-telegram-bot
pytest
websocket-client
